+++ database/migrations/2024_10_03_072026_create_sales_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('products_id');
             $table->string('vendor');
             $table->integer('qty');
-            $table->float('total_harga')->nullable()->default(0);
+            $table->decimal('total_harga',15,2)->nullable()->default(0);
             $table->timestamps();
 
             $table->foreign('products_id')->references('id')->on('products')->onDelete('cascade');
